refactor: change structure and names of some folders and components
