+++ src/app/authorization/page.tsx
@@ -1,5 +1,5 @@
 'use client';
-import SignInForm from '../components/authorization/signIn';
+import SignInForm from '../components/forms/SignInForm';
 
 export default function Authorization() {
   return (
