+++ src/app/components/forms/Forms.css
@@ -1,0 +1,4 @@
+.registration-link {
+  text-decoration: dashed;
+  color: var(--link-color1);
+}
