+++ src/app/components/forms/SignInForm.tsx
@@ -7,7 +7,9 @@ import { UserSchemaWithoutConfirm } from '../utils/yup';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { signInWithEmailAndPassword } from 'firebase/auth';
 import { auth } from '../utils/firebase';
-import { useAppSelector } from '@/app/lib/hooks/redux';
+import Link from 'next/link';
+import { useAppSelector } from '@/app/lib/redux/hooks/redux';
+import './Forms.css';
 
 type authData = {
   email: string;
@@ -34,21 +36,30 @@ const SignInForm = () => {
 
   const onSubmit = async (data: authData) => {
     const { email, password } = data;
-    console.log(data);
+
     signInWithEmailAndPassword(auth, email, password).catch((error) => {
       console.log(error);
     });
   };
+
   if (authUser) {
     return null;
   }
+
   return (
     <form className="submit" onSubmit={handleSubmit(onSubmit)}>
       <input placeholder="Email" {...register('email')} />
       {errors.email && <p className="error">{errors.email.message}</p>}
+
       <input placeholder="Password" type="password" {...register('password')} />
       {errors.password && <p className="error">{errors.password.message}</p>}
+
       <button type="submit">Sign In</button>
+
+      <p>{`Don't have an account yet?`}</p>
+      <Link className="registration-link" href="registration">
+        Registration ❯
+      </Link>
     </form>
   );
 };
