+++ src/app/components/links/AuthPageLink.tsx
@@ -3,13 +3,8 @@ import './Links.css';
 
 export default function AuthPageLink() {
   return (
-    <>
-      <Link href="registration" className="page-link">
-        <span className="link-arrow">Sign Up</span>
-      </Link>
-      <Link href="authorization" className="page-link">
-        <span className="link-arrow">Sign In</span>
-      </Link>
-    </>
+    <Link href="authorization" className="page-link">
+      <span className="link-arrow">Sign In / Sign Up ❯</span>
+    </Link>
   );
 }
