+++ src/app/languages/languages.ts
@@ -9,4 +9,10 @@ export const languages = {
     'Пожалуйста, авторизуйтесь для доступа к приложению!',
     'Будь ласка, авторизуйтесь для доступу до програми!',
   ],
+  welcome: [
+    'Welcome to our app!',
+    'Сардэчна запрашаем у наш дадатак!',
+    'Добро пожаловать в наше приложение!',
+    'Ласкаво просимо до нашого додатку!',
+  ],
 };
