+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import Header from './components/header/Header';
 import Footer from './components/footer/Footer';
-import { LangContextProvider } from './store/langContext';
+import { LangContextProvider } from './lib/context/langContext';
 import StoreProvider from './storeProvider';
 import './globals.css';
