+++ src/app/lib/redux/store.ts
@@ -1,10 +1,12 @@
 'use client';
+
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
 import authReducer from './reducers/auth';
 
 export const rootReducer = combineReducers({
   authReducer,
 });
+
 export const setupStore = () => {
   return configureStore({
     reducer: rootReducer,
