+++ src/app/page.css
@@ -1,3 +1,4 @@
+.main-page-title,
 .auth-note-container {
   display: flex;
   flex-direction: column;
