+++ src/app/registration/page.tsx
@@ -1,5 +1,7 @@
 'use client';
-import SignUpForm from '../components/authorization/signUp';
+
+import SignUpForm from '../components/forms/SignUpForm';
+
 export default function Authorization() {
   return (
     <main className="main">
